Material UI - Notes View
