+++ package.json
@@ -53,6 +53,7 @@
     "express": "^4.17.1",
     "install": "^0.13.0",
     "lodash": "^4.17.15",
+    "moment": "^2.24.0",
     "npm": "^6.13.0",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
