+++ src/pages/Landing.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import {Typography, Container} from '@material-ui/core'
-import NoteSample from './NoteSample'
-/* import NotesContainer from '../components/notes/NotesContainer' */
+import NotesContainer from '../components/notes/NotesContainer'
 
 
 
@@ -9,12 +8,10 @@ const Landing = () => {
     return (
         <React.Fragment>
             <Container maxWidth = 'lg'>
-                <Typography component='div' style={{height: '100vh' }} ><NoteSample /></Typography>
+                <Typography component='div' style={{height: '100vh' }} ><NotesContainer /></Typography>
             </Container>
         </React.Fragment>
     )
 }
 
 export default Landing
-
-/*<div><NotesContainer/></div>*/
