+++ src/redux/constants.js
@@ -1,0 +1,1 @@
+export const NOTES_LOCAL_STORAGE_ID = 'notes'
